# KDesktopVirt/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: KVirtualStage API
  description: |
    Playwright-equivalent desktop automation platform for AI agents.
    
    KVirtualStage provides comprehensive desktop automation capabilities through multiple interfaces:
    - REST API for remote control
    - WebSocket streaming for live desktop viewing
    - Language bindings for Python, Node.js, and C/C++
    - MCP server for AI agent integration
    - CLI and TUI for interactive management
    
    ## Features
    - Natural human-like cursor movement and typing
    - Cross-platform desktop environment support
    - Session recording and playback
    - Workflow automation with multiple steps
    - Real-time performance monitoring
    - Enterprise-grade security and authentication
    
    ## Getting Started
    1. Create a new session with your desired desktop environment
    2. Use automation commands to control the desktop
    3. Optionally start recording to capture interactions
    4. Execute complex workflows with multiple automation steps
    
    ## Authentication
    All API endpoints require valid session authentication. Sessions are created through the `/api/v1/sessions` endpoint.
    
  version: 0.1.0
  contact:
    name: KVirtualStage Team
    url: https://github.com/KooshaPari/KVirtualStage
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.kvirtualstage.dev/v1
...
paths:
  /sessions:
    post:
      summary: Create a new desktop automation session
      description: |
        Creates a new virtual desktop session with the specified configuration.
        The session provides an isolated desktop environment for automation.
      tags:
        - Session Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            examples:
              ubuntu_session:
                summary: Ubuntu desktop session
                value:
                  user_id: "demo_user"
                  session_name: "my_automation_session"
                  desktop_type: "ubuntu"
              kde_session:
                summary: KDE desktop session
                value:
                  user_id: "power_user"
                  session_name: "kde_workflow"
                  desktop_type: "ubuntu-kde"
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  value:
                    success: true
                    data:
                      session_id: "session_20241212_143022"
                      status: "active"
                      vnc_url: "ws://localhost:8080/api/v1/sessions/session_20241212_143022/stream"
                      ws_url: "ws://localhost:8080/api/v1/sessions/session_20241212_143022/stream"
                    error: null
                    timestamp: "2024-12-12T14:30:22Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    get:
      summary: List all active sessions
      description: Retrieves a list of all currently active desktop automation sessions.
      tags:
        - Session Management
      responses:
        '200':
          description: List of active sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                sessions_list:
                  value:
                    success: true
                    data:
                      - session_id: "session_20241212_143022"
                        user_id: "demo_user"
                        desktop_type: "ubuntu"
                        status: "active"
                        recording_active: false
                      - session_id: "session_20241212_144501"
                        user_id: "power_user"
                        desktop_type: "ubuntu-kde"
                        status: "active"
                        recording_active: true
                    error: null
                    timestamp: "2024-12-12T14:45:30Z"
  /sessions/{session_id}:
    get:
      summary: Get session information
      description: Retrieves detailed information about a specific session.
      tags:
        - Session Management
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      summary: Remove/close a session
      description: Terminates and removes a desktop automation session.
      tags:
        - Session Management
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /sessions/{session_id}/cursor/move:
    post:
      summary: Move cursor naturally
      description: |
        Moves the cursor to specified coordinates using natural, human-like movement patterns.
        The movement includes realistic acceleration, deceleration, and subtle variations.
      tags:
        - Automation Control
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveCursorRequest'
            examples:
              center_screen:
                summary: Move to center of screen
                value:
                  target_x: 400
                  target_y: 300
              corner_position:
                summary: Move to corner
                value:
                  target_x: 50
                  target_y: 50
      responses:
        '200':
          description: Cursor moved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /sessions/{session_id}/mouse/click:
    post:
      summary: Click at current or specified position
      description: |
        Performs a mouse click with natural timing and pressure simulation.
        Can click at current cursor position or move to coordinates first.
      tags:
        - Automation Control
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClickRequest'
            examples:
              left_click:
                summary: Left click at current position
                value:
                  button: "left"
              right_click_at_position:
                summary: Right click at specific coordinates
                value:
                  x: 200
                  y: 150
                  button: "right"
      responses:
        '200':
          description: Click executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /sessions/{session_id}/keyboard/type:
    post:
      summary: Type text naturally
      description: |
        Types text with natural human-like timing, including realistic delays,
        occasional corrections, and variable typing speed.
      tags:
        - Automation Control
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TypeTextRequest'
            examples:
              simple_text:
                summary: Type simple message
                value:
                  text: "Hello, World!"
              fast_typing:
                summary: Type at fast speed
                value:
                  text: "This is a longer message for demonstration."
                  wpm: 90
      responses:
        '200':
          description: Text typed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /sessions/{session_id}/recording/start:
    post:
      summary: Start session recording
      description: |
        Begins recording the desktop session to a video file.
        Supports multiple quality levels and formats.
      tags:
        - Recording
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartRecordingRequest'
            examples:
              high_quality:
                summary: High quality recording
                value:
                  output_filename: "demo_session.mp4"
                  quality: "high"
              streaming_quality:
                summary: Streaming optimized
                value:
                  output_filename: "live_demo.mp4"
                  quality: "streaming"
      responses:
        '200':
          description: Recording started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /sessions/{session_id}/recording/stop:
    post:
      summary: Stop session recording
      description: Stops the current recording and finalizes the video file.
      tags:
        - Recording
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Recording stopped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /sessions/{session_id}/workflow:
    post:
      summary: Execute automation workflow
      description: |
        Executes a complex automation workflow consisting of multiple steps.
        Each step can be a cursor movement, click, text input, or wait operation.
      tags:
        - Workflow Automation
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowRequest'
            examples:
              calculator_demo:
                summary: Calculator demonstration
                value:
                  name: "Calculator Demo"
                  description: "Demonstrate basic calculator usage"
                  continue_on_error: false
                  steps:
                    - name: "Open calculator"
                      action_type: "move_cursor"
                      parameters:
                        x: 100
                        y: 100
                      timeout_seconds: 5
                    - name: "Click calculator"
                      action_type: "click"
                      parameters:
                        button: "left"
                      timeout_seconds: 5
                    - name: "Type calculation"
                      action_type: "type"
                      parameters:
                        text: "2 + 2 ="
                      timeout_seconds: 10
      responses:
        '200':
          description: Workflow executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /sessions/{session_id}/stream:
    get:
      summary: WebSocket stream connection
      description: |
        Establishes a WebSocket connection for real-time desktop streaming.
        Provides live video feed and interactive control capabilities.
      tags:
        - Streaming
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '101':
          description: WebSocket connection established
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      description: Unique session identifier
      schema:
        type: string
        example: "session_20241212_143022"
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        data:
          type: object
          description: Response data (varies by endpoint)
          nullable: true
        error:
          type: string
          description: Error message if request failed
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of the response
      required:
        - success
        - timestamp
    CreateSessionRequest:
      type: object
      properties:
        user_id:
          type: string
          description: User identifier for the session
          example: "demo_user"
        session_name:
          type: string
          description: Optional name for the session
          example: "my_automation_session"
        desktop_type:
          type: string
          enum: ["ubuntu", "ubuntu-xfce", "ubuntu-kde", "centos", "fedora", "arch", "debian"]
          description: Desktop environment type
          default: "ubuntu"
          example: "ubuntu"
      required:
        - user_id
    MoveCursorRequest:
      type: object
      properties:
        target_x:
          type: number
          format: float
          description: Target X coordinate
          example: 400.0
        target_y:
          type: number
          format: float
          description: Target Y coordinate
          example: 300.0
      required:
        - target_x
        - target_y
    ClickRequest:
      type: object
      properties:
        x:
          type: number
          format: float
          description: Optional X coordinate (moves cursor first if provided)
          example: 200.0
        y:
          type: number
          format: float
          description: Optional Y coordinate (moves cursor first if provided)
          example: 150.0
        button:
          type: string
          enum: ["left", "right", "middle"]
          description: Mouse button to click
          default: "left"
          example: "left"
    TypeTextRequest:
      type: object
      properties:
        text:
          type: string
          description: Text to type
          example: "Hello, World!"
        wpm:
          type: number
          format: float
          description: Words per minute typing speed
          default: 65.0
          example: 65.0
      required:
        - text
    StartRecordingRequest:
      type: object
      properties:
        output_filename:
          type: string
          description: Output video filename
          example: "demo_session.mp4"
        quality:
          type: string
          enum: ["low", "medium", "high", "streaming"]
          description: Recording quality level
          default: "medium"
          example: "high"
      required:
        - output_filename
    WorkflowRequest:
      type: object
      properties:
        name:
          type: string
          description: Workflow name
          example: "Calculator Demo"
        description:
          type: string
          description: Workflow description
          example: "Demonstrate basic calculator usage"
        continue_on_error:
          type: boolean
          description: Continue execution on step errors
          default: false
        steps:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowStep'
          description: List of workflow steps
      required:
        - name
        - steps
    WorkflowStep:
      type: object
      properties:
        name:
          type: string
          description: Step name
          example: "Click calculator button"
        action_type:
          type: string
          enum: ["move_cursor", "click", "type", "wait"]
          description: Type of action to perform
          example: "click"
        parameters:
          type: object
          description: Action-specific parameters
          example: {"x": 100, "y": 100, "button": "left"}
        timeout_seconds:
          type: number
          format: float
          description: Step timeout in seconds
          default: 30
          example: 5
      required:
        - name
        - action_type
        - parameters
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            invalid_params:
              value:
                success: false
                data: null
                error: "Invalid parameters: user_id is required"
                timestamp: "2024-12-12T14:30:22Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            session_not_found:
              value:
                success: false
                data: null
                error: "Session not found: session_invalid_id"
                timestamp: "2024-12-12T14:30:22Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            server_error:
              value:
                success: false
                data: null
                error: "Internal server error: failed to initialize desktop environment"
                timestamp: "2024-12-12T14:30:22Z"
...